mvc and dao pattern
