+++ EasyBarber/Controller/controller_barbeiro.cpp
@@ -1,0 +1,41 @@
+#include "controller_barbeiro.h"
+
+Controller_Barbeiro::Controller_Barbeiro(){
+    dao_barbeiro = new DAO_Barbeiro();
+}
+
+int Controller_Barbeiro::buscarHorariosAgendados(Barbeiro* barbeiro){
+    vector<Horario*>lista_horarios = dao_barbeiro->buscaHorariosAgendados(barbeiro);
+    int cont = 0;
+    for(Horario* horario: lista_horarios){
+        barbeiro->add(horario);
+        cont++;
+    }
+    return cont;
+}
+
+bool Controller_Barbeiro::updateNome(Barbeiro* barbeiro, const string& nome){
+    if(!dao_barbeiro->updateNome(barbeiro, nome)) return false;
+    barbeiro->setNome(nome);
+    return true;
+}
+bool Controller_Barbeiro::updateEndereco(Barbeiro* barbeiro, const string& endereco){
+    if(!dao_barbeiro->updateEndereco(barbeiro, endereco)) return false;
+    barbeiro->setEndereco(endereco);
+    return true;
+}
+bool Controller_Barbeiro::updateTelefone(Barbeiro* barbeiro, const string& telefone){
+    if(!dao_barbeiro->updateTelefone(barbeiro, telefone)) return false;
+    barbeiro->setTelefone(telefone);
+    return true;
+}
+bool Controller_Barbeiro::updateSenha(Barbeiro* barbeiro, const string& senha){
+    if(!dao_barbeiro->updateSenha(barbeiro, senha)) return false;
+    barbeiro->setSenha(senha);
+    return true;
+}
+bool Controller_Barbeiro::updatePreco(Barbeiro* barbeiro, const float& preco){
+    if(!dao_barbeiro->updatePreco(barbeiro, preco)) return false;
+    barbeiro->setPreco(preco);
+    return true;
+}
