+++ EasyBarber/DAO/dao_pessoa.h
@@ -1,0 +1,14 @@
+#ifndef DAO_PESSOA_H
+#define DAO_PESSOA_H
+
+#include <QtSql>
+#include "../Model/pessoa.h"
+
+class DAO_Pessoa
+{
+public:
+    DAO_Pessoa();
+    virtual vector<Horario*> buscaHorariosAgendados(Pessoa * pessoa, char identifier);
+};
+
+#endif // DAO_PESSOA_H
