+++ EasyBarber/EasyBarber.pro
@@ -9,28 +9,43 @@ CONFIG += c++17
 #DEFINES += QT_DISABLE_DEPRECATED_BEFORE=0x060000    # disables all the APIs deprecated before Qt 6.0.0
 
 SOURCES += \
-    barbearia.cpp \
-    barbeiro.cpp \
-    cliente.cpp \
-    horario.cpp \
+    DAO/dao_barbearia.cpp \
+    DAO/dao_barbeiro.cpp \
+    DAO/dao_cliente.cpp \
+    DAO/dao_pessoa.cpp \
+    Model/barbearia.cpp \
+    Model/barbeiro.cpp \
+    Model/cliente.cpp \
+    Model/horario.cpp \
+    Model/pessoa.cpp \
     main.cpp \
-    pessoa.cpp \
-    telaagendamento.cpp \
-    telaescolhabarbeiro.cpp \
-    telaescolhadata.cpp
+    View/telaagendamento.cpp \
+    View/telaescolhabarbeiro.cpp \
+    View/telaescolhadata.cpp \
+    Controller/controller_barbearia.cpp \
+    Controller/controller_cliente.cpp \
+    Controller/controller_barbeiro.cpp
 
 HEADERS += \
-    barbearia.h \
-    barbeiro.h \
-    cliente.h \
-    horario.h \
-    pessoa.h \
-    telaagendamento.h \
-    telaescolhabarbeiro.h \
-    telaescolhadata.h
+    DAO/dao_barbearia.h \
+    DAO/dao_barbeiro.h \
+    DAO/dao_cliente.h \
+    DAO/dao_pessoa.h \
+    Model/barbearia.h \
+    Model/barbeiro.h \
+    Model/cliente.h \
+    Model/horario.h \
+    Model/pessoa.h \
+    View/telaagendamento.h \
+    View/telaescolhabarbeiro.h \
+    View/telaescolhadata.h \
+    Controller/controller_barbearia.h \
+    Controller/controller_cliente.h \
+    Controller/controller_barbeiro.h
 
 FORMS += \
-    telaescolhadata.ui
+    View/telaescolhadata.ui \
+    View/telaescolhadata.ui
 
 # Default rules for deployment.
 qnx: target.path = /tmp/$${TARGET}/bin
