+++ EasyBarber/Model/Cliente.cpp
@@ -1,0 +1,7 @@
+#include "Cliente.h"
+#include <iostream>
+
+Cliente::Cliente(string nome, string endereco, string telefone, string email, string senha)
+        : Pessoa(nome, endereco, telefone, email, senha) {}
+
+Cliente::~Cliente() {}
